requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
google-auth==2.25.2
aiohttp==3.9.1
python-dotenv==1.0.0
